Alphaedge_Copy/requirements.txt
# AlphaCopy - Copy Trading Framework
# pip install -r requirements.txt

# ── Core async ────────────────────────────────────────────────────────────
aiohttp>=3.9.0
asyncio
websockets>=12.0

# ── Hyperliquid ───────────────────────────────────────────────────────────
hyperliquid-python-sdk>=0.9.0
eth-account>=0.10.0

# ── Binance ───────────────────────────────────────────────────────────────
python-binance>=1.0.19

# ── Polymarket ────────────────────────────────────────────────────────────
py_clob_client>=0.17.0
web3>=6.0.0

# ── Config & utilities ────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Data & analytics ──────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Optional: monitoring ─────────────────────────────────────────────────
# prometheus-client>=0.18.0   # Uncomment for Prometheus metrics
